Add Abstract Form For Common Functions
Add Absturact Form for the common functions between the main form and
Fieldset Nested Form since they share too many functions that have the
same codebase.
