+++ src/components/fieldset_nested_form.js
@@ -1,26 +1,32 @@
 import React from "react"
-import Form from "./form.js"
+import AbstractForm from "./abstract_form.js"
 
 // Nested forms (forms inside nested fieldsets)
-export default class FieldsetNestedForm extends Form {
+export default class FieldsetNestedForm extends AbstractForm {
   displayName = "Frig.FieldsetNestedForm"
 
+  static propTypes = AbstractForm.propTypes
+  static defaultProps = AbstractForm.defaultProps
+
   static childContextTypes = {
+    ...AbstractForm.childContextTypes,
     frigFieldset: React.PropTypes.object,
   }
 
   getChildContext() {
     return {
+      ...super.getChildContext(),
       frigFieldset: {
         index: this.props.index,
       },
     }
   }
 
+
   render() {
     return (
       <div>
-        {this.props.children}
+        { this.props.children }
       </div>
     )
   }
